# Construct nzw in fortran order to reduce cache misses in hot loops.